requirements.txt
django==5.2.9
whitenoise
gunicorn
crispy-bootstrap4
django-crispy-forms
psycopg2-binary
pillow
django-storages
azure-storage-blob